[cherry-pick 2.8] Support transformer 3 (#5655)
* [host] add gather_tree (#5419)
* [host] add gather_nd (#5435)
* [Support transformer] Fix dtype error for compare, expand, fill_constant_batch_size_like , gather_nd and reduce_all ops (#5588)
* fix compare, expand, fill_constant_batch_size_like and gather_nd op
* fix gather_nd, reduce_all ops

+++ lite/kernels/host/crf_decoding_compute.h
@@ -69,7 +69,7 @@ void Decode(const Tensor& emission_weights,
 
   T max_score = -(std::numeric_limits<T>::max)();
   int max_i = 0;
-  for (size_t i = 0; i < tag_num; ++i) {
+  for (int64_t i = 0; i < tag_num; ++i) {
     T score = alpha_value[(seq_len - 1) * tag_num + i] + w[tag_num + i];
     if (score > max_score) {
       max_score = score;
